componentize discussion backlog
